requirements.txt
alembic==1.13.1
blinker==1.7.0
cached-property==1.5.2   
cachelib==0.9.0
certifi==2024.2.2        
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Caching==2.3.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
mypy-extensions==1.0.0
packaging==24.0
pluggy==1.4.0
psycopg2-binary
pytest==8.1.1
python-dotenv==1.0.1
redis==5.0.4
requests==2.31.0
saga_py==1.1.0
SQLAlchemy==2.0.28
sqlalchemy-orm==1.2.10
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
gunicorn==22.0.0